separated app.js into multiple files
